http://bugs.osportfolio.org/jira/browse/OSP-1310
Fixes this issue by attaching more appropriate HTTP headers to the response.  The headers that are sent are: Content-Type (application/octet-stream), Content-Disposition (attachment;filename="abc.xyz"), and Content-Length.  The Transfer-Encoding header is not applied.

DownloadableManager.packageForDownload() now returns a String, which should be the name of the file.

FileDownloadServlet uses this string in the Content-Disposition header for the filename attribute.  If the filename is null or empty, the attribute is omitted, allowing the user agent to determine the filename based on the request URL.

The PresentationManagerImpl is the only implementation that returns filenames in this commit.  The others should be adjusted to return appropriate filenames.


NOTE: There may be a call to use some MIME magic to determine a better Content-Type, but using application/octet-stream usually results in a "Save As" dialog, which should be appropriate for most downloads through the FileDownloadServlet.


git-svn-id: https://source.sakaiproject.org/svn/metaobj/trunk@9932 66ffb92e-73f9-0310-93c1-f5514f145a0a